accumulator

+++ src/Foundational/accumulator/accumulator.h
@@ -17,7 +17,7 @@ template <typename T, typename SUMMER>
 class Accumulator_Base
 {
   protected:
-    unsigned int _n;
+    uint64_t _n;
     T _minval;
     T _maxval;
     SUMMER _xsum;
@@ -33,7 +33,7 @@ class Accumulator_Base
 
     int ok () const;
 
-    unsigned int n () const { return _n;}
+    uint64_t n () const { return _n;}
 
     bool empty() const {
       return _n == 0;
